requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
google-adk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
pytest
